requirements.txt
pandas==1.1.3
opencv-python==4.1.2.30
tqdm==4.54.1
future==0.18.2
Pillow==8.0.1